test main builder
