+++ tools/cell_census_builder/tests/conftest.py
@@ -1,3 +1,4 @@
+import io
 import os
 import pathlib
 from typing import List
@@ -120,6 +121,23 @@ def datasets(assets_path: str) -> List[Dataset]:
             )
 
     return datasets
+
+
+@pytest.fixture
+def manifest_csv(tmp_path: pathlib.Path) -> io.TextIOWrapper:
+    manifest_content = """
+    dataset_id_1, data/h5ads/dataset_id_1.h5ad
+    dataset_id_2, data/h5ads/dataset_id_2.h5ad
+    """
+    path = f"{tmp_path}/manifest.csv"
+    h5ad_path = f"{tmp_path}/data/h5ads/"
+    pathlib.Path(h5ad_path).mkdir(parents=True, exist_ok=True)
+    pathlib.Path("data/h5ads/dataset_id_1.h5ad").touch()
+    pathlib.Path("data/h5ads/dataset_id_2.h5ad").touch()
+    with open(path, "w+") as f:
+        f.writelines(manifest_content.strip())
+
+    return open(path)
 
 
 @pytest.fixture()
